download_gbif_data_from_doi

+++ R/download_gbif_data_from_doi.R
@@ -8,6 +8,7 @@
 #' @param folder Save folder
 #' @param subfolder Save in subfolder
 #' @param keep_only_occurrence_file Keep only occurrence.txt file
+#' @param overwrite overwrite files
 #'
 #' @details Download GBIF occurrence data downloaded from DOI
 #'
@@ -36,7 +37,8 @@
 #' @export
 download_gbif_data_from_doi <- function(gbif_doi_url,
                                         folder = '',
-                                        keep_only_occurrence_file = TRUE)
+                                        keep_only_occurrence_file = TRUE,
+                                        overwrite = TRUE)
 {
 
   if(!dir.exists(folder))
@@ -56,23 +58,28 @@ download_gbif_data_from_doi <- function(gbif_doi_url,
   url_file_zip_GBIF <- li_dr_txt[41]
   path_file_zip_GBIF <- paste0(folder,'\\dataGBIF.zip')
 
-  downloader::download(url_file_zip_GBIF, path_file_zip_GBIF, mode = "wb")
-
-  utils::unzip(path_file_zip_GBIF, exdir = folder)
-
-  if(keep_only_occurrence_file == TRUE)
+  if((!file.exists(url_file_zip_GBIF)) |
+     overwrite == TRUE)
   {
-    # files_tmp <- list.files(path =  folder, full.names = TRUE, recursive = TRUE)
-    # ind_files <- grepl(paste0('occurrence.txt|',family_name,'.zip'), files_tmp)
-    # file.remove( list.files(path =  paste0(folder,'\\dataset'), full.names = TRUE)  )
-    # file.remove(files_tmp[ind_files==FALSE], recursive = TRUE)
+    downloader::download(url_file_zip_GBIF, path_file_zip_GBIF, mode = "wb")
 
-    files_tmp <- list.files(path =  folder, full.names = TRUE)
-    ind_files <- grepl(paste0('occurrence.txt|','dataGBIF.zip'), files_tmp)
-    # file.remove(files_tmp[ind_files==FALSE])
+    utils::unzip(path_file_zip_GBIF, exdir = folder)
 
-    # Delete Dir
-    unlink(files_tmp[ind_files==FALSE],recursive = TRUE)
+    if(keep_only_occurrence_file == TRUE)
+    {
+      # files_tmp <- list.files(path =  folder, full.names = TRUE, recursive = TRUE)
+      # ind_files <- grepl(paste0('occurrence.txt|',family_name,'.zip'), files_tmp)
+      # file.remove( list.files(path =  paste0(folder,'\\dataset'), full.names = TRUE)  )
+      # file.remove(files_tmp[ind_files==FALSE], recursive = TRUE)
+
+      files_tmp <- list.files(path =  folder, full.names = TRUE)
+      ind_files <- grepl(paste0('occurrence.txt|','dataGBIF.zip'), files_tmp)
+      # file.remove(files_tmp[ind_files==FALSE])
+
+      # Delete Dir
+      unlink(files_tmp[ind_files==FALSE],recursive = TRUE)
+    }
+
   }
 
   files_tmp <- list.files(path =  folder, full.names = TRUE, recursive = TRUE)
